Refactor explicit all different constraint
This allows us to implement `boxed_clone`.
